Revert autoSave edits, leave QRF refactored
