+++ src/components/QuestionnaireResponseForm/types.ts
@@ -17,5 +17,4 @@ export interface QRFProps extends QuestionnaireResponseFormProps {
     FormFooterComponent?: React.ElementType<FormFooterComponentProps>;
     saveButtonTitle?: React.ReactNode;
     cancelButtonTitle?: React.ReactNode;
-    autoSave?: boolean;
 }
