+++ src/containers/PatientDetails/PatientDocument/index.tsx
@@ -23,7 +23,6 @@ export interface PatientDocumentProps {
     questionnaireId?: string;
     encounterId?: string;
     onSuccess?: (resource: QuestionnaireResponseFormSaveResponse) => void;
-    autoSave?: boolean;
 }
 
 export function PatientDocument(props: PatientDocumentProps) {
@@ -63,7 +62,7 @@ export function PatientDocument(props: PatientDocumentProps) {
                                 }}
                                 onCancel={() => navigate(-1)}
                                 saveButtonTitle={'Complete'}
-                                autoSave={props.autoSave ? props.autoSave : !provenance}
+                                autoSave={!provenance}
                                 draftSaveResponse={draftSaveResponse}
                                 setDraftSaveResponse={setDraftSaveResponse}
                             />
