+++ src/uberComponents/ResourceListPage/actions.tsx
@@ -5,8 +5,7 @@ import { omit } from 'lodash';
 import { useNavigate } from 'react-router-dom';
 
 import { ModalTrigger } from 'src/components/ModalTrigger';
-import { QuestionnaireResponseForm } from 'src/components/QuestionnaireResponseForm';
-import { QRFProps } from 'src/components/QuestionnaireResponseForm/types';
+import { QRFProps, QuestionnaireResponseForm } from 'src/components/QuestionnaireResponseForm';
 import { questionnaireIdLoader } from 'src/hooks/questionnaire-response-form-data';
 
 import { S } from './styles';
